50/182 errors

+++ srcs/flags/ft_printf_llo.c
@@ -6,7 +6,7 @@ int ft_printf_llo(va_list ap, int display, int tronc)
 	int len;
 
 	len = 0;
-	nb = va_arg(ap, long long);
+	nb = (long long int)va_arg(ap, long long int);
 	len = ft_putnbr_base_printf(nb, "01234567", display);
     return (len);
